+++ src/Integrations/logs/lokiRecordingRules.ts
@@ -5,7 +5,7 @@ import { type SyntaxNode } from '@lezer/common';
 import { lastValueFrom } from 'rxjs';
 
 import { createMetricsLogsConnector, type FoundLokiDataSource } from './base';
-import { findHealthyLokiDataSources } from '../../RelatedLogs/RelatedLogsManager';
+import { findHealthyLokiDataSources } from '../../RelatedLogs/RelatedLogsOrchestrator';
 
 export interface RecordingRuleGroup {
   name: string;
